+++ src/app/modules/login/singin/singin.component.html
@@ -1,6 +1,6 @@
-<mat-card>
+<mat-card *ngIf= "(loginService.switchLogIn)">
   <mat-card-content>
-    <form [formGroup]="loginForm" (ngSubmit)="onLogin()">
+    <form [formGroup]="loginService.loginForm" (ngSubmit)="loginService.onLogin()">
       <h2>Sing In</h2>
       <mat-form-field>
         <input matInput placeholder="Email" name="email" required formControlName="email">
@@ -8,9 +8,9 @@
       <mat-form-field>
         <input matInput type="password" placeholder="Password" name="password" required formControlName="password" >
       </mat-form-field>
-      <button type="submit" mat-raised-button color="primary" (click)="activeLogin()">Login</button>
+      <button type="submit" mat-raised-button color="primary" (click)="loginService.activeLogin()">Login</button>
       <h3>¿No tienes una cuenta?</h3>
-      <button type="submit" mat-raised-button color="primary">Regístrate Aquí</button>
+      <button type="submit" mat-raised-button color="primary" (click)="(loginService.valorCaseLogin = 2)">Regístrate Aquí</button>
     </form>
   </mat-card-content>
 </mat-card>
