+++ src/app/modules/login/singin/singin.component.ts
@@ -9,38 +9,6 @@ import { LoginService } from '../services/login.service';
 })
 export class SinginComponent {
 
-  constructor (private loginService: LoginService, private formBuilder: FormBuilder) {}
-
-  loginForm = this.formBuilder.group({
-    email: new FormControl(''),
-    password: new FormControl('')
-  });
-
-  ngOnInit(): void {
-    this.defaultLogin()
-  }
-
-  defaultLogin() {
-    if(!localStorage.getItem('key'))
-    localStorage.setItem('key', 'false')
-  }
-
-  activeLogin(){
-    localStorage.setItem('key', 'true')
-  }
-
-
-  onLogin(){
-    if( this.loginForm.controls.email.value && this.loginForm.controls.password.value){
-    let form = {
-      email: this.loginForm.controls.email.value,
-      password: this.loginForm.controls.password.value
-    }
-    this.loginService.login(form).subscribe(data =>{
-      console.log(data);
-      this.activeLogin()
-    })
-  }
-  }
+  constructor (public loginService: LoginService) {}
 
 }
