+++ src/app/modules/login/singup/singup.component.scss
@@ -1,7 +1,8 @@
 mat-card {
-  max-width: 400px;
+  max-width: 500px;
   margin: 2em auto;
   text-align: center;
+  padding-bottom: 5%;
 }
 
 mat-form-field {
