+++ src/app/modules/login/singup/singup.component.ts
@@ -1,4 +1,5 @@
 import { Component } from '@angular/core';
+import { LoginService } from '../services/login.service';
 
 @Component({
   selector: 'app-singup',
@@ -6,5 +7,7 @@ import { Component } from '@angular/core';
   styleUrls: ['./singup.component.scss']
 })
 export class SingupComponent {
+
+  constructor (public loginService: LoginService) {}
 
 }
