+++ src/app/modules/shared/home/home.component.html
@@ -1,14 +1,4 @@
 <div class="components">
-<app-singin *ngIf= "true"></app-singin>
-<app-singup *ngIf= "false"></app-singup>
+<app-singin></app-singin>
+<app-singup></app-singup>
 </div>
-
-<!-- <ng-container *ngIf="!mostrar; then mostrar1 else mostrar2"></ng-container>
-
-<ng-template #mostrar1>
-  <app-singin></app-singin>
-</ng-template>
-
-<ng-template #mostrar2>
-  <app-singup></app-singup>
-</ng-template> -->
